+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - is.codion.common.utilities split from is.codion.common.core.
 - is.codion.plugin.jul.proxy, is.codion.plugin.log4j.proxy, is.codion.plugin.logback.proxy META-INF service definition filename fixed.
 - LoggerProxy.getLogLevel() and setLogLevel() overloaded with logger name.
+- LoggerProxy.loggers() added.
 ### is.codion.tools.jul.classpath
 - ClasspathConfiguration split from is.codion.plugin.jul.proxy module.
 ### is.codion.tools.swing.robot
@@ -24,6 +25,7 @@ Codion Change Log
 - ComponentBuilder.ancestorListener() added.
 - ComponentBuilder.hierarchyListener() added.
 - ImagePanel.resetView() renamed reset().
+- LogLevelPanel added.
 ### is.codion.framework.domain
 - ValueAttributeDefinition added, related refactoring.
 ### is.codion.framework.db.core
@@ -39,6 +41,8 @@ Codion Change Log
 ### is.codion.framework.server
 - is.codion.framework.server module now requires is.codion.tools.jul.classpath.
 - is.codion.framework.server module transitive requires for is.codion.framework.db.rmi removed.
+### is.codion.swing.framework.ui
+- EntityApplicationPanel, advanced logging configuration added.
 
 ## 0.18.57
 ### is.codion.common.core